+++ src/lib/components/searchBox.svelte
@@ -4,15 +4,16 @@
   import { fade, slide } from 'svelte/transition';
   import { Fzf } from '../../../node_modules/fzf';
   import type { NameWithFeature } from '../data/features';
-  import { genHoverName, type HoverName, type NameWithFeatures } from '../store';
+  import { HoverSelect } from '../data/searchBox';
+  import type { NameWithFeatures } from '../store';
 
   type Name = NameWithFeature;
   let fzf: [string | undefined, Fzf<readonly string[]>][];
 
   export let overlayFilter: string;
   export let names: NameWithFeatures[];
-  export let curr: HoverName<Name>;
-  curr = genHoverName<Name>({});
+  export let curr: HoverSelect<Name>;
+  curr = new HoverSelect<Name>();
 
   let showSearch = true;
